fix: remove user session prop drilling

+++ actions/delete-url.ts
@@ -1,10 +1,16 @@
 "use server";
 
 import { LinkTable } from "@/data-access/urls";
+import { getSession } from "@/lib/get-session";
 
-export const deleteURL = async (id: number, userId: string) => {
+export const deleteURL = async (id: number) => {
+
+  const session = await getSession();
+  if (!session) return { error: "not authorized" };
+  const { id: uid } = session.user;
+
   const response = await LinkTable.deleteLinkById({
-    userId,
+    userId: uid,
     id
   });
